step 3: first attempt at writing the getOrElse macro, type error in expansion
Since we can't seem to implement getOrElse in a zero-overhead fashion using
conventional means, let's write a macro for that.

NB! It is the general consensus that macros should only be used as a last resort,
i.e. when non-macro-based techniques either can't achieve the necessary effect
or become prohibitively verbose or awkward. We don't recommend using macros
just for fun, because macros are typically harder to write, understand and maintain
that vanilla Scala. Check out our "Macros vs Types" talk for examples of this trade-off:
http://scalamacros.org/paperstalks/2014-03-01-MacrosVsTypes.pdf.

    import scala.language.experimental.macros
    import scala.reflect.macros.blackbox.Context

    final class Optional[+A >: Null](val value: A) extends AnyVal {
      def get: A = value
      def isEmpty = value == null
      def getOrElse[B >: A](alt: => B): B = macro OptionalMacros.getOrElse
      override def toString = if (isEmpty) "<empty>" else s"$value"
    }

    class OptionalMacros(val c: Context) {
      def getOrElse(alt: c.Tree): c.Tree = {
        import c.universe._
        q"if (isEmpty) $alt else value"
      }
    }

The sketch shown here utilizes the usual macro-def-refers-to-macro-impl scheme
that might be familiar to you from Scala 2.10, but it also makes use of
two neat new features introduced in Scala 2.11:

1) Macro bundles. In Scala 2.10, only static objects could define macro
implementations, which then had to define a `c: Context` parameter and
be path-dependent on it, which caused complications when macros grew big
enough to warrant modularization. In Scala 2.11, the context can come
from the enclosing class, which greatly simplifies moving macro logic around.

2) Quasiquotes. In Scala 2.10, the only high-level way of constructing
trees was `c.universe.reify`, which only works if the types of the trees
it constructs are known ahead of time. This has proven to be a major
inconvenience, so in Scala 2.11 we implemented quasiquotes, which are
arenot only more flexible for tree construction, but also support
high-level syntax for pattern matching. Quasiquotes are now
the recommended way to perform tree manipulations.

The code that the macro needs to generate is very simple, and
our macro implementation looks correct. However if we try to do `sbt run`
to test out the macro, we'll get a compilation error. Let's try to
figure this out.

[error] /Users/xeno_by/Projects/macrology201/core/src/main/scala/Test.scala:6: not found: value isEmpty
[error]     val x1 = x.getOrElse(new C)
[error]                         ^
[error] /Users/xeno_by/Projects/macrology201/core/src/main/scala/Test.scala:10: not found: value isEmpty
[error]     val y1 = y.getOrElse({ println("hello"); new C })
[error]                         ^
[error] two errors found