step 8: the MatchError hints that we forgot to match type args of the application
After we fix the error, things start working!

[info] Running Test
(C,C)
hello
(<empty>,C)

In cases when MatchErrors are more involved, first try to do showCode,
because toString-based MatchError messages might be misleading. If that
doesn't help, go for showRaw and try to construct a non-quasiquote based
pattern match - it might be a bug in quasiquotes.