step 11: let's try to hack our macro by inducing a name clash
This and the subsequent couple dozen commits will showcase typical macro
robustness issues and techniques to safeguard against these pitfalls.

We are going to begin with name clashes - the number one source
of trickiness in the macroland.

A peculiar thing about macros is that they fuse two lexical scopes
that are disparate in human-written source code of the program:
the namespace of a macro implementation and the namespace of
a macro expansion callsite. This can lead to tricky inadvertent
name clashes that can manifest themselves only under very rare
circumstances.

Let's try to hack our macro a bit. We know that the macro expansion
introduces a temporary variable called `temp`, so let's create
our own variable `temp` at the macro expansion site and try to
refer to it in a macro argument:

    val temp = 100
    println(new Optional(if (temp < 100) new C else null).getOrElse(new C))

Here's how the expansion is going to look like. Looks like we're going
to run into a compilation error, right?

    def main(args: Array[String]): Unit = {
      val temp: Int = 100;
      scala.this.Predef.println(({
        val temp: Optional[C] = new Optional[C](if (temp.<(100))
          new C()
        else
          null);
        if (temp.isEmpty)
          new C()
        else
          temp.value
      }: C))
    }

Wrong! Interestingly enough everything works fine, even though the inner variable
`temp` that we're generating should have shadowed the outer variable `temp`
that's referred by the prefix of the macro expansion.