step 12: -uniqid explains the phenomenon we're observing
Trees constituting the macro application are pretypechecked before being
passing into macros, so they remember the variables that they refer to
(unless someone does `c.resetAllAttrs` - the API that was present in Scala 2.10,
but was removed in Scala 2.11).

    def main#15596(args#15598: Array#880[String#7353]): Unit#2602 = {
      val temp#15811: Int#1078 = 100;
      scala#27.this.Predef#1715.println#7577(({
        val temp#17557: Optional#7797[C#7806] = new Optional#7797[C#7806](if (temp#15811.<#15871(100))
          new C#7806()
        else
          null);
        if (temp#17557.isEmpty#16045)
          new C#7806()
        else
          temp#17557.value#16039
      }: C#7806))
    }

We can see that internally the compiler is able to discern the two `temp` variables
by assigning them with unique identifiers called symbols (symbol ids is what gets
printed after hash signs when -uniqid is enabled).

That way the `temp` in `new Optional(if (temp < 100) new C else null)` can
remember that it refers to the `temp` at the macro expansion site, regardless of whether
there are intervening `temp` definitions generated by a macro.