step 0: genesis and motivation
Let's write a simple implicit materializer that will do deep immutability
checking for us. This can be useful in cases where we'd like to
statically enforce that no mutable state is passed around in concurrent
setting:

    trait ActorApi {
      def send[T: Immutable](msg: T): Unit
    }

Here we'd like to specify that our actor API only supports sending of
messages that are considered to be immutable via a simple type class:

    trait Immutable[T]

Whenever an instance of this type class is available we consider this
T to be immutable by contract. Providing instances for all the
immutable things can be quite a tedious task and we'd like not to do it
manually but rather delegate this job to a macro.