step 21: a non-solution: c.untypecheck (né c.resetAttrs)
An easy workaround to the owner chain corruption problem is erasing the
chunk of the symbol table whose corresponding code is undergoing radical
transformations. After the symbols in the problematic trees are erased,
post-expansion typecheck is going to recreate them from scratch, keeping
the symbol table consistent. This is done by a dedicated macro API called
c.untypecheck in Scala 2.11 and c.resetLocalAttrs in Scala 2.10.