step 22: an example that breaks c.untypecheck
The untypecheck workaround would be great if it worked robustly, but
unfortunately that's not the case.

Due to internal implementation details of scalac's typechecker, some tree
shapes can't be typechecked twice, which means that typechecking them
once, resetting attrs and then typechecking again isn't going to work:
https://issues.scala-lang.org/browse/SI-5464.

One of the easy ways to break c.untypecheck is to use a lazy val.
Here's what happens if we go for it:

    [error] /Users/xeno_by/Projects/macrology201/core/src/main/scala/Test.scala:6: missing parameter type
    [error]     println(foo({ lazy val y = new Optional(null); y }).getOrElse(new C))
    [error]                            ^
    [error] one error found

Therefore we need to look for better ways of dealing with our problem.
Unfortunately, in general case the alternatives are all much more complicated
than c.untypecheck. That's why in practice people sometimes choose to live
with c.untypecheck's known issues.