step 26: an inlining transformer
Deep tree rewritings are usually performed by subclassing Transformer
(there's also Traverser that can be used to visit trees).

In our case, we need to replace all references to the argument of the lambda
with references to a temporary variable that holds the prefix of the
macro application. In order to do that, we check all Ident tree nodes, find those
that are bound to the argument of the lambda and rewrite them into q"$temp.value".