step 1: a dummy materializer
As a zero-th approximation we provide evidence that every type is
immutable. The resulting macro is very simple, but even it demonstrates
a few important points:

1. We choose whitebox context because we'd like the compiler to continue
searching for implicit instances even if our macro fails. Usage of
blackbox context would have implied that we can provide an
evidence no matter what, so the failure of a macro expansion
becomes a compilation error.

2. Our type parameter needs a weak typetag so that we can reflect on it
in a macro via weakTypeOf. Here we just use this type to create a dummy
type class instance, but later on we'll be performing meaningful introspective
operations on it.