step 2: testing that the dummy materializer is wrong
Even though it's clear that our dummy materializer is going to
wrongly produce `Immutable` instances for any time, including
mutable ones, we need a test to expose the wrong behavior so
that our future versions can fix it for good.

To test whether T is immutable we're adding a utility method called `is`
that does a compile-time check for whether Immutable[T] is available
and returns a runtime boolean.

The `is` method is implemented as a simple macro that performs a search
for an implicit value and returns true if it finds something and false
if it doesn't. We don't actually need to explicitly branch here as we can
just unquote boolean expression into a quasiquote and it will convert
it to a tree for us (via lifting).

Note how we create a type for Immutable[T], having only a type for T at hand.
First of all, tq"..." isn't going to cut it, because it returns a Tree
not a Type, and c.inferImplicitValue needs a Type. Secondly, Scala doesn't
have quasiquotes for types yet, so we have to assemble the type by hand.
In general this requires c.typecheck(..., c.TYPEmode), but for this particular
case there's a shortcut.