step 3: reflecting the members
To fix our test case we reflect the members of the type and check
any of those are mutable fields. In case if such fields exist,
we discard such type as mutable and abort with an error message.

Since our implicit macro is whitebox, the error message will not
be normally shown if we interact with our implicit materializer
through implicit search but can see it if we enable the -Xlog-implicits
flag or call Immutable.materialize directly.