step 4: deep immutability
Not only do we need to check that fields of the T are vals but we also
need to check that types of vals are themeselves immutable. To check
this we collect types of all fields and than produce calls to `implicitly`
that ensure that if our macro expansion compiles that means that all
fields are immutable.