step 5: tying the knot
Our previous attempt at transitive immutability has one pitfall: it
would stack overflow the compiler on recursive classes.

To fix this problem we introduce implicit object of our current type
early in the expansion block so that calls to `implicitly` will not expand
materializer on the same type twice, but will rather resolve to the implicit
object that we generated.