step 7: open classes vs final classes
Another aspect our materializer doesn't take care of is subtyping: our
type might not have any mutable fields but at runtime it might turn out
to be a subtype that added some mutable fields.

To guard against such case we require the class to be final and fail
with error for any open classes.

This means that closures (subtypes of FunctionN classes, which are open)
will be considered mutable. That's correct, because closures can capture
arbitrary state - immutable and mutable alike, but also too restrictive.
Scala spores, which by the way also use macros, are an ongoing research
into immutable closures.