step 15: how to deal with variance
An interesting thing about materializers was uncovered just the day before
the workshop.

If a materializee has a covariant type parameter, then T will always be
Nothing. If it's invariant, then it'll be Any. Looks like a nasty artifact
of how implicit macros integrate into type inference, but actually there's
a simple way of dealing with it.

Providing a materializer with an dummy invariant context bound fixes
variant type parameters in place, preventing them from gravitating towards
the lower or upper bound.