step 16: a bit of philosophy
Should Immutable[T] be invariant or contravariant? (It definitely can't
be covariant, because if a class is immutable, it doesn't guarantee that
its superclasses, e.g. Any, are immutable).

That's an interesting question, so let's discuss, but in the meanwhile
I'm going to revert the previous commit to guaranteedly stay sound.